/* Suppose you have a table users with columns first_name, last_name, and date_of_birth. If you want to create a unique key using all these columns, you can use the FARM_FINGERPRINT function, which is a hash function provided by BigQuery: */
sql
Copy code
SELECT
  first_name,
  last_name,
  date_of_birth,
  FARM_FINGERPRINT(CONCAT(first_name, last_name, CAST(date_of_birth AS STRING))) AS unique_key
FROM
  `your_dataset.users`;
Here:

/*
We are using the CONCAT function to concatenate first_name, last_name, and date_of_birth.
CAST(date_of_birth AS STRING) is used to ensure that the date of birth (which is likely a DATE type) is treated as a string for the concatenation.
FARM_FINGERPRINT then computes the hash of the concatenated string to produce the unique_key.
This will give you a unique key for each combination of the three columns. However, do keep in mind that while hash collisions (different inputs producing the same hash) are extremely rare with good hash functions, they are theoretically possible. If you're using this approach for any critical application where absolute uniqueness is required, make sure to have checks in place to handle potential collisions.
*/